move Content.Source impl to receiver subclass
